uint32 changed to uint64 to prevent collisions
