+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/gurnoorsinghbhinder/brevo-assignments
+
+go 1.24.1
