+++ main.go
@@ -1,0 +1,8 @@
+package main
+
+import HashFunction "github.com/gurnoorsinghbhinder/brevo-assignments/hashfunction"
+
+func main(){
+	HashFunction.HashFunction()
+	
+}
